Move SafeAlternationTests to PatternInternalTest to avoid breaking crosscheck
